+++ tests/bash/run_pebble_dev.sh
@@ -1,0 +1,3 @@
+docker run --network=host --rm -e PEBBLE_WFE_NONCEREJECT=0 \
+--name pebble \
+ghcr.io/letsencrypt/pebble:2.6.0